AIR-QUALITY/AQ-UI-With-Mail/requirements.txt
geopy==2.4.1
python-dotenv==1.1.1
requests==2.32.4
streamlit==1.48.0
